Guard route using a protected route
